requirements.txt
dvc
mlflow
pandas
numpy
scikit-learn
joblib
pyyaml
python-dotenv
requests
fastapi
uvicorn
pydantic
